requirements.txt
Jinja2==3.1.2
MarkupSafe==2.1.1
pytidylib==0.3.2
six==1.16.0
